cmd: rearrange cmds
